added background anc score title and game loop

+++ Engine/Object/Tilemap.cpp
@@ -4,7 +4,7 @@ namespace jc
 {
 	void Tilemap::Create()
 	{
-		for (size_t i = 0; i < tiles.size(); i++)
+		for (int i = 0; i < tiles.size(); i++)
 		{
 			int index = tiles[i];
 			if (index == 0) continue;
